Move complex relationship discussion to separate page
Add section on executing raw SQL

Add info about commit() and rollback()
