--- pcweb/pages/docs/database/__init__.py
@@ -1,3 +1,0 @@
-from .overview import database_overview
-from .queries import queries
-from .tables import tables
