--- pcweb/pages/docs/database/queries.py
@@ -1,7 +1,0 @@
-from pcweb import flexdown
-from pcweb.templates.docpage import docpage
-
-
-@docpage()
-def queries():
-    return flexdown.render_file("docs/database/queries.md")
