--- pcweb/pages/docs/database/tables.py
@@ -1,7 +1,0 @@
-from pcweb import flexdown
-from pcweb.templates.docpage import docpage
-
-
-@docpage()
-def tables():
-    return flexdown.render_file("docs/database/tables.md")
